✅ test(flags): Update ListenPortFlag test to use uint type
Changed the test assertion for ListenPortFlag to compare against a uint type instead of uint64. This aligns the test with the recent refactor of flags to use uint, ensuring consistency and correctness in type usage across the codebase.

+++ internal/cli/shared/flags_test.go
@@ -63,7 +63,7 @@ func TestListenPortFlag(t *testing.T) {
 	var flag = shared.ListenPortFlag
 
 	assert.Equal(t, "port", flag.Name)
-	assert.Equal(t, uint64(8080), flag.Value)
+	assert.Equal(t, uint(8080), flag.Value)
 	assert.Contains(t, flag.Sources.String(), "LISTEN_PORT")
 
 	for giveValue, wantErrMsg := range map[uint]string{
